Remove parrots support
