+++ mmengine/optim/optimizer/default_constructor.py
@@ -11,7 +11,7 @@ from mmengine.registry import (OPTIM_WRAPPER_CONSTRUCTORS, OPTIM_WRAPPERS,
                                OPTIMIZERS)
 from mmengine.utils import is_list_of
 from mmengine.utils.dl_utils import mmcv_full_available
-from mmengine.utils.dl_utils.parrots_wrapper import _BatchNorm, _InstanceNorm
+from mmengine.utils.dl_utils.torch_wrapper import _BatchNorm, _InstanceNorm
 from .optimizer_wrapper import OptimWrapper
