+++ mmengine/utils/dl_utils/__init__.py
@@ -3,7 +3,7 @@
 from .collect_env import collect_env
 from .hub import load_url
 from .misc import has_batch_norm, is_norm, mmcv_full_available, tensor2imgs
-from .parrots_wrapper import TORCH_VERSION
+from .torch_wrapper import TORCH_VERSION
 from .setup_env import set_multi_processing
 from .time_counter import TimeCounter
 from .torch_ops import torch_meshgrid
