+++ mmengine/utils/dl_utils/torch_ops.py
@@ -2,11 +2,10 @@
 import torch
 
 from ..version_utils import digit_version
-from .parrots_wrapper import TORCH_VERSION
+from .torch_wrapper import TORCH_VERSION
 
 _torch_version_meshgrid_indexing = (
-    'parrots' not in TORCH_VERSION
-    and digit_version(TORCH_VERSION) >= digit_version('1.10.0a0'))
+    digit_version(TORCH_VERSION) >= digit_version('1.10.0a0'))
 
 
 def torch_meshgrid(*tensors):
